CS 204 – Advanced Programming-  HW3
In this homework, I am going to implement a two-dimensional hybrid linked list (2DHLL) structure that will be used to store information about employees and their assigned projects. Along with the 2DHLL structure, I will also implement and use a stack structure to keep track of the changes that are done on the 2DHLL structure for possible undoings.